+++ slpn_miner/slpn_opt_entropic_relevance_discovery.py
@@ -6,10 +6,10 @@ import scipy
 
 from pm4py.objects.log.importer.xes import importer as xes_importer
 
-from src.slpn_visualiser import visualize_slpn, view
-from src.symbolic_conversion import calculate_inverse_poland_expression, get_inverse_poland_expression
-from src.util import setup, get_slpn
-from src.slpn_exporter import export_slpn, export_slpn_xml
+from slpn_miner.slpn_visualiser import visualize_slpn, view
+from slpn_miner.symbolic_conversion import calculate_inverse_poland_expression, get_inverse_poland_expression
+from slpn_miner.util import setup, get_slpn
+from slpn_miner.slpn_exporter import export_slpn, export_slpn_xml
 
 
 def optimize_with_er(log, pn, im, fm):
