+++ slpn_miner/stochastic_equation_system.py
