+++ slpn_miner/symbolic_conversion.py
