+++ slpn_miner/util.py
@@ -4,10 +4,10 @@ import logging
 from pm4py.objects.log.obj import EventLog, Trace
 from pm4py.algo.filtering.log.variants import variants_filter
 
-from src.log_util import get_stochastic_language
-from src.stochastic_cross_product import ConstructCP
-from src.stochastic_reachability_graph import construct_stochastic_reachability_graph
-from src.trace_dfa import create_dfa_from_list
+from slpn_miner.log_util import get_stochastic_language
+from slpn_miner.stochastic_cross_product import ConstructCP
+from slpn_miner.stochastic_reachability_graph import construct_stochastic_reachability_graph
+from slpn_miner.trace_dfa import create_dfa_from_list
 
 
 logging.getLogger().setLevel(logging.DEBUG)
